Made two hovercraft prefabs
I added two hovercraft prefabs for our project, attached with "invisible" boxes. They're not actually invisible, but you can make them so by removing the material from the Mesh Renderer of the boxes. I got the behavior I wanted, with the crafts at a height of 7. I recommend only removing the rendering during play and leaving it during development.

One ISSUE I noticed is that the turn speed and the movement speed of the Movement script are tied together. Changing the turn speed has no effect, but changing the movement speed changes the turn speed, too.